externalize result codes, change project structure
