+++ app/src/main/java/gr/thmmy/mthmmy/utils/parsing/NewParseTask.java
@@ -2,6 +2,8 @@ package gr.thmmy.mthmmy.utils.parsing;
 
 import org.jsoup.nodes.Document;
 
+import gr.thmmy.mthmmy.utils.NetworkTask;
+
 public abstract class NewParseTask<T> extends NetworkTask<T> {
 
     public NewParseTask(OnParseTaskStartedListener onParseTaskStartedListener, OnParseTaskCancelledListener onParseTaskCancelledListener,
